add osrf_pycommon.vendor.trollius_helper
This package will add an embedded version of
trollius to the PYTHONPATH if trollius is not
found in the system.

It will not always have the embedded version of
trollius, in which cases it will just fail if
trollius is not found in the system.

Currently the only target case for embedding
trollius is when making the Debian binary with
stdeb. This is done because there is no trollius
binary for Debian/Ubuntu.

The setup.py file has new logic which embeds
trollius 2.1 when building a .deb. This is
also avoided for Python versions < 3.4, since
asyncio is available for all of these.