Suppress warning for specifically handled behavior
In Python 3.10, there is a deprecation warning raised when calling
get_event_loop when there is no running loop. In the future, it will
raise a RuntimeError, which is already handled by this code in exactly
the way we want it to.

Once we drop support for python < 3.8, we can change this function to
get_running_loop and drop the warning filter.