Catch all of the spurious warnings from get_event_loop.
The previous patch that did this only did it for one of
the cases, but we actually need to catch this warning for
all cases of get_event_loop().  Do that here by wrapping
the entire block in catch_warnings().  With this in
place, other spurious warnings from downstream packages
(like launch) are also quieted.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>